+++ iexec-blockchain-adapter-api-library/src/main/java/com/iexec/blockchain/api/CommandStatus.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 IEXEC BLOCKCHAIN TECH
+ * Copyright 2022-2023 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.iexec.blockchain.tool;
+package com.iexec.blockchain.api;
 
-public enum Status {
+public enum CommandStatus {
     RECEIVED,
     PROCESSING,
     SUCCESS,