+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 - Add and use a non-root user in the dockerfile. (#113)
 - Reindent `build.gradle` file. (#114)
 - Standardisation of the dockerfile and its location in regard to other java components. (#115)
+- Rename `Status` to `CommandStatus` in library. (#117)
 
 ### Dependency Upgrades